Move getVariableUsesById and getTopBlocks to utils file, so that it can be used by both devtools and cleanup-plus
